requirements.txt
Flask
Flask-Session
Flask-Cors
python-dotenv
openai>=1.0.0
gunicorn
